Extract CurrentBlockDeviceInfo to its own file
Rename SyncLVAndLVDL to SyncPVAndLVDL
